+++ polylx/__init__.py
@@ -4,6 +4,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 import pandas as pd
 import networkx as nx
+import seaborn as sns
 from .core import Grain, Boundary, Grains, Boundaries, Sample
 from .utils import deg, circular
 
@@ -11,4 +12,4 @@ __author__ = 'Ondrej Lexa'
 __email__ = 'lexa.ondrej@gmail.com'
 
 __all__ = ['Grain', 'Boundary', 'Grains', 'Boundaries', 'Sample',
-           'np', 'plt', 'pd', 'nx', 'deg', 'circular']
+           'np', 'plt', 'pd', 'nx', 'sns', 'deg', 'circular']